Seperate test out into module file
